Create visual odometer class
